requirements.txt
beautifulsoup4~=4.9.3
python-dateutil~=2.8.1
python-decouple~=3.4
selenium~=3.141.0
